requirements.txt
asgiref==3.10.0
click==8.3.1
colorama==0.4.6
cssbeautifier==1.15.4
dj-database-url==3.0.1
Django==5.2.8
django-filter==25.2
djangorestframework==3.16.1
djlint==1.36.4
EditorConfig==0.17.1
jsbeautifier==1.15.4
json5==0.12.1
Markdown==3.10
pathspec==0.12.1
python-decouple==3.8
PyYAML==6.0.3
regex==2025.11.3
ruff==0.14.6
six==1.17.0
sqlparse==0.5.3
tomli==2.3.0
tqdm==4.67.1
typing_extensions==4.15.0
whitenoise==6.11.0
